Cloud-Engineering/nu-labs/423-yaml-lab-activity/interactive/requirements.txt
pandas~=1.4.0
pyyaml~=6.0